feat: link emu with arm
