+++ go.mod
@@ -29,3 +29,5 @@ require (
 	google.golang.org/protobuf v1.28.1
 	gopkg.in/yaml.v3 v3.0.1
 )
+
+replace github.com/cloudwego/frugal => /Users/bytedance/go/src/github.com/cloudwego/frugal
